Changed ints to double (#77)

+++ be/src/Project/Services/Basket/Basket.Api/IntegrationEventsModels/BasketCheckoutProductEventModel.cs
@@ -5,6 +5,6 @@ namespace Basket.Api.IntegrationEventsModels
     public class BasketCheckoutProductEventModel
     {
         public Guid? ProductId { get; set; }
-        public int BookedQuantity { get; set; }
+        public double BookedQuantity { get; set; }
     }
 }
