+++ be/src/Project/Web/Buyer/Buyer.Web/Areas/Products/ApiResponseModels/ProductStockResponseModel.cs
@@ -5,7 +5,7 @@ namespace Buyer.Web.Areas.Products.ApiResponseModels
 {
     public class ProductStockResponseModel : BaseResponseModel
     {
-        public int? AvailableQuantity { get; set; }
+        public double? AvailableQuantity { get; set; }
         public int? RestockableInDays { get; set; }
         public DateTime? ExpectedDelivery { get; set; }
     }
